glog: Bring back removed functions
This restores the functions removed in commit 37cd6274. The reason to
bring them back is to be compatible with github.com/golang/glog. If we
are API compatible then it is possible to use a go.mod replace
directive to replace any dependency's use of github.com/golang/glog
with github.com/aristanetworks/glog. This resolves a pain point of
trying to use both glog libraries in one binary.

I believe the advantages from this change outweigh the negative
stylistic concerns of an API that contains redundant or unwanted
functionality. For example, Info and Infoln do the same thing and the
Warning level being an ill-defined in-between of Info and Error.

Change-Id: Ifcfcf5e4fa07e1a16b3ebac72a9ab646f7e1e3f1